pyproject.toml
[tool.poetry]
name = "agents101"
version = "0.1.0"
description = "LLM agent creation 101"
authors = [
    "Yann MEUROU <ymeurou2702@gmail.com>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
langchain-mistralai = "^0.2.10"
python-dotenv = "^1.1.0"
faiss-cpu = "^1.11.0"
mistralai = "^1.8.1"
pypdf = "^5.6.0"
pypdf2 = "^3.0.1"
langchain = "^0.3.25"
streamlit = "^1.40.0"
streamlit-chat = "^0.2.0"
streamlit-extras = "^0.5.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"